Make mapsearch bookmarkable.
The only real problem I have remaining on this is that a bookmark
which does not contain full lat/long/zoom coordinates will not
decluster and will present with a selected cluster/location at the
default map position.

To fix this will require a day or two of refactoring to eliminate
the current doubling up on models in the Layout.

Fixes #1